Updated, Fixed & Reorganized Project
Please note that some of the included Pods, despite being up-to-date, still had restricted optimization for Xcode 10 and Swift 3. I made some changes to the frameworks directly to be compatible with Xcode 11.3 and Swift 5.

BUILD NOTES
• Converted project to Swift 5
• Updated project for Xcode 11.3
• Updated Pods and related frameworks
• Updated & converted frameworks to Swift 5
• Added additional code to frameworks for stability
• Fixed theme name update delay
• Minor bug fixes
• Resolved errors & warnings

macOS
• Updated for macOS 10.15 (Catalina)
• Updated for Xcode 11.3
• Converted to Swift 5

Furthermore, in the near future, I wish to create a more extensive example project using this framework. Either a file manager or a git client.